+++ lib/src/specs/field.g.dart
@@ -12,43 +12,33 @@ class _$Field extends Field {
   @override
   final BuiltList<Object> docs;
   @override
-  final Code assignment;
+  final Code? assignment;
   @override
   final bool static;
   @override
   final String name;
   @override
-  final Reference type;
+  final Reference? type;
   @override
   final FieldModifier modifier;
 
-  factory _$Field([void Function(FieldBuilder) updates]) =>
+  factory _$Field([void Function(FieldBuilder)? updates]) =>
       (new FieldBuilder()..update(updates)).build() as _$Field;
 
   _$Field._(
-      {this.annotations,
-      this.docs,
+      {required this.annotations,
+      required this.docs,
       this.assignment,
-      this.static,
-      this.name,
+      required this.static,
+      required this.name,
       this.type,
-      this.modifier})
+      required this.modifier})
       : super._() {
-    if (annotations == null) {
-      throw new BuiltValueNullFieldError('Field', 'annotations');
-    }
-    if (docs == null) {
-      throw new BuiltValueNullFieldError('Field', 'docs');
-    }
-    if (static == null) {
-      throw new BuiltValueNullFieldError('Field', 'static');
-    }
-    if (name == null) {
-      throw new BuiltValueNullFieldError('Field', 'name');
-    }
-    if (modifier == null) {
-      throw new BuiltValueNullFieldError('Field', 'modifier');
-    }
+    BuiltValueNullFieldError.checkNotNull(annotations, 'Field', 'annotations');
+    BuiltValueNullFieldError.checkNotNull(docs, 'Field', 'docs');
+    BuiltValueNullFieldError.checkNotNull(static, 'Field', 'static');
+    BuiltValueNullFieldError.checkNotNull(name, 'Field', 'name');
+    BuiltValueNullFieldError.checkNotNull(modifier, 'Field', 'modifier');
   }
 
   @override
@@ -100,12 +90,12 @@ class _$Field extends Field {
 }
 
 class _$FieldBuilder extends FieldBuilder {
-  _$Field _$v;
+  _$Field? _$v;
 
   @override
   ListBuilder<Expression> get annotations {
     _$this;
-    return super.annotations ??= new ListBuilder<Expression>();
+    return super.annotations;
   }
 
   @override
@@ -117,7 +107,7 @@ class _$FieldBuilder extends FieldBuilder {
   @override
   ListBuilder<Object> get docs {
     _$this;
-    return super.docs ??= new ListBuilder<Object>();
+    return super.docs;
   }
 
   @override
@@ -127,13 +117,13 @@ class _$FieldBuilder extends FieldBuilder {
   }
 
   @override
-  Code get assignment {
+  Code? get assignment {
     _$this;
     return super.assignment;
   }
 
   @override
-  set assignment(Code assignment) {
+  set assignment(Code? assignment) {
     _$this;
     super.assignment = assignment;
   }
@@ -151,25 +141,25 @@ class _$FieldBuilder extends FieldBuilder {
   }
 
   @override
-  String get name {
+  String? get name {
     _$this;
     return super.name;
   }
 
   @override
-  set name(String name) {
+  set name(String? name) {
     _$this;
     super.name = name;
   }
 
   @override
-  Reference get type {
+  Reference? get type {
     _$this;
     return super.type;
   }
 
   @override
-  set type(Reference type) {
+  set type(Reference? type) {
     _$this;
     super.type = type;
   }
@@ -189,14 +179,15 @@ class _$FieldBuilder extends FieldBuilder {
   _$FieldBuilder() : super._();
 
   FieldBuilder get _$this {
-    if (_$v != null) {
-      super.annotations = _$v.annotations?.toBuilder();
-      super.docs = _$v.docs?.toBuilder();
-      super.assignment = _$v.assignment;
-      super.static = _$v.static;
-      super.name = _$v.name;
-      super.type = _$v.type;
-      super.modifier = _$v.modifier;
+    final $v = _$v;
+    if ($v != null) {
+      super.annotations = $v.annotations.toBuilder();
+      super.docs = $v.docs.toBuilder();
+      super.assignment = $v.assignment;
+      super.static = $v.static;
+      super.name = $v.name;
+      super.type = $v.type;
+      super.modifier = $v.modifier;
       _$v = null;
     }
     return this;
@@ -204,14 +195,12 @@ class _$FieldBuilder extends FieldBuilder {
 
   @override
   void replace(Field other) {
-    if (other == null) {
-      throw new ArgumentError.notNull('other');
-    }
+    ArgumentError.checkNotNull(other, 'other');
     _$v = other as _$Field;
   }
 
   @override
-  void update(void Function(FieldBuilder) updates) {
+  void update(void Function(FieldBuilder)? updates) {
     if (updates != null) updates(this);
   }
 
@@ -224,12 +213,15 @@ class _$FieldBuilder extends FieldBuilder {
               annotations: annotations.build(),
               docs: docs.build(),
               assignment: assignment,
-              static: static,
-              name: name,
+              static: BuiltValueNullFieldError.checkNotNull(
+                  static, 'Field', 'static'),
+              name:
+                  BuiltValueNullFieldError.checkNotNull(name, 'Field', 'name'),
               type: type,
-              modifier: modifier);
+              modifier: BuiltValueNullFieldError.checkNotNull(
+                  modifier, 'Field', 'modifier'));
     } catch (_) {
-      String _$failedField;
+      late String _$failedField;
       try {
         _$failedField = 'annotations';
         annotations.build();