+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## 4.0.1-dev
+
+## 4.0.0
+
+* Migrate to null safety.
+* Changed the DartEmittor constructor to use named optional parameters.
+* Add `ParenthesizedExpression` and
+  `ExpressionVisitor.visitParenthesizedExpression.`
+
+## 3.7.0
+
+* Add support for converting a Method to a generic closure, with
+  `Method.genericClosure`.
+
 ## 3.6.0
 
 * Add support for creating `extension` methods.